# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionInput": {
        "properties": {
          "evidence_url": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Evidence Url",
            "type": "string"
          },
          "original_file": {
            "default": false,
            "title": "Original File",
            "type": "boolean"
          },
          "provider_transfer": {
            "default": false,
            "title": "Provider Transfer",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "sensitivity_cleared": {
            "default": false,
            "title": "Sensitivity Cleared",
            "type": "boolean"
          },
          "user_text": {
            "default": false,
            "title": "User Text",
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "evidence_url"
        ],
        "title": "DecisionInput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetadataInput": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expected_version": {
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagView"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "expected_version",
          "reason",
          "description",
          "tags"
        ],
        "title": "MetadataInput",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CultureChamp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/revisions/{revision_id}": {
      "get": {
        "operationId": "admin_revision_api_v1_admin_revisions__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/approve": {
      "post": {
        "operationId": "approve_revision_api_v1_admin_revisions__revision_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/metadata": {
      "patch": {
        "operationId": "amend_revision_metadata_api_v1_admin_revisions__revision_id__metadata_patch",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Amend Revision Metadata",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/metadata-history": {
      "get": {
        "operationId": "revision_metadata_history_api_v1_admin_revisions__revision_id__metadata_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MetadataEventView"
                  },
                  "title": "Response Revision Metadata History Api V1 Admin Revisions  Revision Id  Metadata History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision Metadata History",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/original": {
      "get": {
        "operationId": "admin_original_api_v1_admin_revisions__revision_id__original_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Original",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/retry": {
      "post": {
        "operationId": "retry_revision_api_v1_admin_revisions__revision_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/revoke": {
      "post": {
        "operationId": "revoke_revision_api_v1_admin_revisions__revision_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/segment-history": {
      "get": {
        "operationId": "revision_segment_history_api_v1_admin_revisions__revision_id__segment_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SegmentReviewEventView"
                  },
                  "title": "Response Revision Segment History Api V1 Admin Revisions  Revision Id  Segment History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision Segment History",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/segments": {
      "patch": {
        "operationId": "review_revision_segments_api_v1_admin_revisions__revision_id__segments_patch",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentReviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Revision Segments",
        "tags": [
          "admin-sources"
        ]
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "operationId": "materials_list_api_v1_materials_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "people",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "People"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "application/pdf",
                    "text/plain",
                    "text/csv",
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Media Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaterialView"
                  },
                  "title": "Response Materials List Api V1 Materials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Materials List",
        "tags": [
          "materials"
        ]
      }
    },
...
    "/api/v1/materials/{revision_id}": {
      "get": {
        "operationId": "material_detail_api_v1_materials__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Material Detail",
        "tags": [
          "materials"
        ]
      }
    },
...
  }
}